agentify-webshop/requirements.txt
a2a-sdk
google-generativeai
python-dotenv
uvicorn
fastapi
httpx
gym==0.26.2
spacy==3.8.11
thefuzz
cleantext
unidecode
flask
selenium
pyserini==0.17.0
torch
numpy<2
tomli
pillow
bs4
lxml
html5lib
requests
